requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
litellm>=1.17.0
pydantic>=2.7.0
python-dotenv>=1.0.0
loguru>=0.7.0 